TEST(day_10): add go_forward tests (#4)
* TEST(day_10): add go_forward tests

The goal is to force implementation of function go_forward, which could
be used iteratively later to count number of steps to the furthest pipe.

First argument is a reference to the pipe_maze.
Second argument of go_forward is current_pipe.
Third argument is passed as support to let developer to determine the
direction of further step.

First scenario basically check, whether we move from '-' to the opposite
direction then we came from i.e. here opposite to previous pipe 'S'.
go_forward() should give back coordinates of '7' and then following '|'
In second scenario it should give back coordinates of 'L' and then
following '-'.

Used map:
.....
.S-7.
.|.|.
.L-J.
.....

* REFAC(day_10): rename go_forward uts